+++ src/Ydb.Sdk.OpenTelemetry/src/TracerProviderBuilderExtensions.cs
@@ -8,9 +8,16 @@ namespace Ydb.Sdk.OpenTelemetry;
 public static class TracerProviderBuilderExtensions
 {
     /// <summary>
-    /// Subscribes to the Ydb.Sdk activity source to enable ADO.NET tracing.
+    /// Subscribes to all Ydb.Sdk activity sources.
     /// </summary>
-    public static TracerProviderBuilder AddYdb(this TracerProviderBuilder builder) => builder.AddSource("Ydb.Sdk");
+    public static TracerProviderBuilder AddYdb(this TracerProviderBuilder builder) =>
+        builder.AddYdbAdo().AddYdbTopic();
+
+    /// <summary>
+    /// Subscribes to the Ydb.Sdk.Ado activity source to enable ADO.NET tracing.
+    /// </summary>
+    public static TracerProviderBuilder AddYdbAdo(this TracerProviderBuilder builder) =>
+        builder.AddSource("Ydb.Sdk.Ado");
 
     /// <summary>
     /// Subscribes to the Ydb.Sdk.Topic activity source to enable Topic tracing.